Refactored sandbox to use actual code files. Added more features to LoggingBehaviourOptions.cs and EventLoggingBehaviour.cs Moved some diagram features to the DiagramService.cs

+++ Blazored.Diagrams/Behaviours/MoveBehaviour.cs
@@ -82,8 +82,8 @@ public class MoveBehaviour : BaseBehaviour
 
         // Apply inverse zoom to the movement
         var inverseZoom = 1 / obj.Model.Zoom;
-        var xDiffAdjusted = xDiff * (double)inverseZoom;
-        var yDiffAdjusted = yDiff * (double)inverseZoom;
+        var xDiffAdjusted = xDiff * inverseZoom;
+        var yDiffAdjusted = yDiff * inverseZoom;
 
         SelectedModels.ForEach(model =>
         {
